Use different <testcase name=…> if using argument
To ensure that tools that parse the generated JUnit would consider those as different test cases, instead of multiple runs of the same test case.

We already have <property name="argument" value="…"> to help us distinguish, but in practice most of the tools that read JUnit XML files will now know about that property or account for it, and only look at the `<testcase classname=…name=…>` attributes to differentiate cases.